% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{cran_version}
\alias{cran_version}
\title{check whether cran version is greater than stated}
\usage{
cran_version(package, version, note)
}
\arguments{
\item{package}{package name}

\item{version}{version to compare}

\item{note}{explain yourself}
}
\value{
the version number of the package on cran, invisibly
}
\description{
check whether cran version is greater than stated
}
