% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{issue_closed}
\alias{issue_closed}
\title{trigger error if an issue is closed}
\usage{
issue_closed(repo, note)
}
\arguments{
\item{repo}{a github repository, in remotes shorthand}

\item{note}{explain yourself}
}
\description{
trigger error if an issue is closed
}
