% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy-tag.R
\name{roxy_tag_parse.roxy_tag_todoordie}
\alias{roxy_tag_parse.roxy_tag_todoordie}
\title{Roxygen tags for todoordie}
\usage{
\method{roxy_tag_parse}{roxy_tag_todoordie}(x)
}
\arguments{
\item{x}{a tag}
}
\value{
something
}
\description{
Roxygen tags for todoordie
}
