% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after_date.R
\name{after_date}
\alias{after_date}
\title{trigger error if after a certain date}
\usage{
after_date(year, month, day)
}
\arguments{
\item{year}{year}

\item{month}{month}

\item{day}{day}
}
\value{
the current date, invisibly
}
\description{
trigger error if after a certain date
}
