% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after_date.R
\name{after_date}
\alias{after_date}
\title{trigger error if after a certain date}
\usage{
after_date(due_at, note)
}
\arguments{
\item{due_at}{an ISO-8601 formatted date}

\item{note}{explain yourself}
}
\value{
the current date, invisibly
}
\description{
trigger error if after a certain date
}
