% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{issue_closed}
\alias{issue_closed}
\title{trigger error if an issue is closed}
\usage{
issue_closed(note, repo)
}
\arguments{
\item{note}{explain yourself}

\item{repo}{a repository, in}
}
\description{
trigger error if an issue is closed
}
