% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{pr_merged}
\alias{pr_merged}
\title{trigger error if a PR is merged}
\usage{
pr_merged(note, repo)
}
\arguments{
\item{note}{explain yourself}

\item{repo}{a github repository, in remotes shorthand}
}
\description{
trigger error if a PR is merged
}
